refactor and test file reading utils
