+++ seqr/utils/file_utils.py
@@ -1,38 +1,57 @@
 import logging
 import os
+import subprocess
 
-from seqr.utils.gcloud.google_bucket_file_utils import is_google_bucket_file_path, \
-    does_google_bucket_file_exist, google_bucket_file_iter
 
 logger = logging.getLogger(__name__)
 
 
-def is_local_file_path(file_path):
-    return "://" not in file_path
+def _run_gsutil_command(command, gs_path, gunzip=False):
+    #  Anvil buckets are requester-pays and we bill them to the anvil project
+    project_arg = '-u anvil-datastorage ' if gs_path.startswith('gs://fc-secure') else ''
+    command = 'gsutil {project_arg}{command} {gs_path}'.format(
+        project_arg=project_arg, command=command, gs_path=gs_path,
+    )
+    if gunzip:
+        command += " | gunzip -c -q - "
+
+    logger.info('==> {}'.format(command))
+    return subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, shell=True)
+
+
+def _is_google_bucket_file_path(file_path):
+    return file_path.startswith("gs://")
 
 
 def does_file_exist(file_path):
-    if is_google_bucket_file_path(file_path):
-        return does_google_bucket_file_exist(file_path)
-    elif is_local_file_path(file_path):
-        return os.path.isfile(file_path)
-    else:
-        raise ValueError("This type of file path is not supported: %(file_path)" % locals())
+    if _is_google_bucket_file_path(file_path):
+        process = _run_gsutil_command('ls', file_path)
+        return process.wait() == 0
+    return os.path.isfile(file_path)
 
 
 def file_iter(file_path, byte_range=None):
-    if is_google_bucket_file_path(file_path):
-        for line in google_bucket_file_iter(file_path, byte_range=byte_range):
+    if _is_google_bucket_file_path(file_path):
+        for line in _google_bucket_file_iter(file_path, byte_range=byte_range):
             yield line
-    elif is_local_file_path(file_path):
+    else:
         with open(file_path) as f:
             if byte_range:
                 f.seek(byte_range[0])
-                while f.tell() < byte_range[1]:
-                    for line in f:
+                for line in f:
+                    if f.tell() < byte_range[1]:
                         yield line
+                    else:
+                        break
             else:
                 for line in f:
                     yield line
-    else:
-        raise ValueError("This type of file path is not supported: %(file_path)" % locals())
+
+
+def _google_bucket_file_iter(gs_path, byte_range=None):
+    """Iterate over lines in the given file"""
+    range_arg = ' -r {}-{}'.format(byte_range[0], byte_range[1]) if byte_range else ''
+    process = _run_gsutil_command('cat{}'.format(range_arg), gs_path, gunzip=gs_path.endswith("gz"))
+    for line in process.stdout:
+        yield line
+
